Move dialog data interface
