+++ src/app/app-components/dialogs/confirm-dialog/confirm-dialog.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { ConfirmDialogData } from '../confirm-dialog-data';
+import { ConfirmDialogData } from './confirm-dialog-data';
 
 @Component({
   selector: 'app-confirm-dialog',
